% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher.R
\name{launch}
\alias{launch}
\title{Launch the ShinyMetabase app}
\usage{
launch(data)
}
\arguments{
\item{data}{A \link{\code{mSet}} object. If is parsed, the data will be loaded
when the ShinyMetabase app is lunched.}
}
\description{
Launch the ShinyMetabase app
}
\examples{
ShinyMetabase::launch(Metabase::lipid)
ShinyMetabase::launch()
}
